+++ NEWS.md
@@ -1,6 +1,9 @@
 NEWS
 ===========
 
+# eixport 0.4.1 (Release date: 22 nov 2019)
+- improves wrf_put, argument mult now can have any length supporting different hourly profiles
+
 # eixport 0.4.0 (Release date: 14 nov 2019)
 - adds sfx_explode to split lines with https://github.com/hypertidy/silicate/issues/102
 - adds as-raster in wrf_grid